refactor: restore original file locations for metadata and initrd_state
Without feature gates there is no need to move files. Restore:
- cmd/boot_service/metadata.rs to original location (with METADATA_PATH_IN_INITRD added)
- cmd/boot_service/initrd_state.rs to original location
- config/initrd_state.rs back to importing from cmd::boot_service::initrd_state
- Delete cmd/metadata.rs (was the moved copy)
- Update import in disk/kernel.rs back to cmd::boot_service::metadata
